feat: class notifications added
