+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.dscvit.vitty">
 
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.WAKE_LOCK" />
 
     <application
         android:name=".App"
@@ -56,15 +57,9 @@
             android:name=".activity.ScheduleActivity"
             android:screenOrientation="portrait" />
 
-        <receiver android:name=".notif.AlarmReceiver" />
-
-        <service
-            android:name=".service.MyFirebaseMessagingService"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
+        <receiver
+            android:name=".notif.AlarmReceiver"
+            android:enabled="true" />
 
         <service
             android:name=".service.TodayWidgetService"
