+++ app/src/main/res/drawable/ic_notif.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="73dp"
+    android:height="30dp"
+    android:viewportWidth="73"
+    android:viewportHeight="30">
+  <path
+      android:pathData="M22.7247,6.7202L18.3345,0H72.3337L67.3582,6.7202H58.1388L43.5049,27.6275H34.2855L48.3341,6.7202H22.7247Z"
+      android:fillColor="#126AFC"/>
+  <path
+      android:pathData="M10.2063,2.5117H0C3.0619,6.9121 18.142,27.652 19.6051,28.9688C21.0681,30.2855 24.4951,30.8209 26.9738,27.594L38.9297,11.0192H29.4524L22.4538,20.7L10.2063,2.5117Z"
+      android:fillColor="#126AFC"/>
+</vector>
